Add count of ignition orgs created
[#156228689]

This will show the total count of orgs with the ignition org quota assigned to them, aka total number of orgs created via ignition. Since the call to get the total number of orgs with the assigned quota can take several seconds, we async update the count in the background every 1m (by default).